+++ R/genBankReadIF.R
@@ -20,14 +20,14 @@ genBankReadIF <- function(filename.gb) {
 
   value <- field <- NULL
 
-  requireNamespace("dplyr")
-
-  if(requireNamespace("tidyr")==FALSE){
-    message(crayon::red("You need to install tidyr"))
-    return(NULL)
-  } else {
-    requireNamespace("tidyr")
-  }
+  # requireNamespace("dplyr")
+  #
+  # if(requireNamespace("tidyr")==FALSE){
+  #   message(crayon::red("You need to install tidyr"))
+  #   return(NULL)
+  # } else {
+  #   requireNamespace("tidyr")
+  # }
 
   if( nchar(filename.gb)>300 ){
     # filename.gb<-nostoc_seqs